requirements.txt
pandas>=1.5
numpy>=1.24
backtesting>=0.3.8
google-cloud-storage>=2.10.0
matplotlib>=3.6
python-binance>=1.0.16   # optional if you extend downloader to fetch from Binance
